More improvements to incremental build
