+++ CMakeLists.txt
@@ -33,26 +33,55 @@ add_compile_options("$<$<CXX_COMPILER_ID:MSVC>:/utf-8>")
 
 list(APPEND ALICE_SOURCES_LIST
 	"src/main.cpp"
-	"src/map/map_state.cpp"
-	"src/map/map_data_loading.cpp"
-	"src/map/map_borders.cpp"
-	"src/map/map.cpp"
-	"src/parsing/parsers.cpp"
-	"src/parsing/float_from_chars.cpp"
-	"src/text/text.cpp"
-	"src/text/fonts.cpp"
-	"src/graphics/texture.cpp"
-	"src/common_types/date_interface.cpp"
 	"src/ai/ai.cpp"
+	"src/common_types/blake2.cpp"
+	"src/common_types/date_interface.cpp"
+	"src/common_types/prng.cpp"
 	"src/culture/culture.cpp"
+	"src/culture/politics.cpp"
+	"src/culture/rebels.cpp"
+	"src/economy/demographics.cpp"
 	"src/economy/economy.cpp"
+	"src/gamestate/commands.cpp"
+	"src/gamestate/diplomatic_messages.cpp"
 	"src/gamestate/modifiers.cpp"
+	"src/gamestate/notifications.cpp"
 	"src/gamestate/serialization.cpp"
+	"src/graphics/opengl_wrapper.cpp"
+	"src/graphics/texture.cpp"
+	"src/gui/gui_common_elements.cpp"
+	"src/gui/gui_effect_tooltips.cpp"
+	"src/gui/gui_graphics.cpp"
+	"src/gui/gui_leader_select.cpp"
+	"src/gui/gui_leader_tooltip.cpp"
+	"src/gui/gui_modifier_tooltips.cpp"
+	"src/gui/gui_province_window.cpp"
+	"src/gui/gui_trigger_tooltips.cpp"
+	"src/gui/map_tooltip.cpp"
+	"src/gui/topbar_subwindows/gui_population_window.cpp"
+	"src/gui/topbar_subwindows/gui_production_window.cpp"
+	"src/gui/topbar_subwindows/gui_technology_window.cpp"
+	"src/gui/topbar_subwindows/gui_budget_window.cpp"
+	"src/gui/unit_tooltip.cpp"
+	"src/map/map.cpp"
+	"src/map/map_borders.cpp"
+	"src/map/map_data_loading.cpp"
+	"src/map/map_modes.cpp"
+	"src/map/map_state.cpp"
 	"src/military/military.cpp"
 	"src/nations/nations.cpp"
+	"src/network/network.cpp"
+	"src/parsing/float_from_chars.cpp"
+	"src/parsing/parsers.cpp"
+	"src/platform_specific.cpp"
 	"src/provinces/province.cpp"
 	"src/scripting/effects.cpp"
+	"src/scripting/events.cpp"
 	"src/scripting/triggers.cpp"
+	"src/text/bmfont.cpp"
+	"src/text/fonts.cpp"
+	"src/text/text.cpp"
+	"src/zstd/zstd.cpp"
 )
 
 if(WIN32)
