+++ src/common_types/blake2.cpp
@@ -1,0 +1,3 @@
+extern "C" {
+#include "blake2.c"
+};
