+++ src/entry_point_nix.cpp
@@ -1,3 +1,7 @@
+#ifdef LOCAL_USER_SETTINGS
+#include "local_user_settings.hpp"
+#endif
+#include "serialization.hpp"
 #include "system_state.hpp"
 
 static sys::state game_state; // too big for the stack
