+++ src/gui/gui_province_window.hpp
@@ -1905,9 +1905,4 @@ public:
 	friend class province_national_focus_button;
 };
 
-void province_national_focus_button::button_action(sys::state& state) noexcept {
-	auto province_window = static_cast<province_view_window*>(state.ui_state.province_window);
-	province_window->nf_win->set_visible(state, !province_window->nf_win->is_visible());
-}
-
 } // namespace ui
