+++ src/platform_specific.cpp
@@ -1,0 +1,30 @@
+#ifdef _WIN64
+// WINDOWS implementations go here
+
+#pragma comment(lib, "d2d1.lib")
+#pragma comment(lib, "D3D11.lib")
+#pragma comment(lib, "dxguid.lib")
+#pragma comment(lib, "d3dcompiler.lib")
+
+#include "simple_fs_win.cpp"
+#include "window_win.cpp"
+#include "sound_win.cpp"
+#include "opengl_wrapper_win.cpp"
+
+#ifndef ALICE_NO_ENTRY_POINT
+#include "entry_point_win.cpp"
+#endif
+
+#else
+// LINUX implementations go here
+
+#include "simple_fs_nix.cpp"
+#include "window_nix.cpp"
+#include "sound_nix.cpp"
+#include "opengl_wrapper_nix.cpp"
+
+#ifndef ALICE_NO_ENTRY_POINT
+#include "entry_point_nix.cpp"
+#endif
+
+#endif
