+++ src/scripting/events.cpp
@@ -1,4 +1,7 @@
 #include "events.hpp"
+#include "effects.hpp"
+#include "gui_event.hpp"
+#include "prng.hpp"
 #include "system_state.hpp"
 #include "triggers.hpp"
